user can now manage listings; final commit for now

+++ app/Http/Controllers/ListingController.php
@@ -45,6 +45,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -57,6 +59,12 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing){
+        
+        // Make sure logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -79,7 +87,21 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing){
+
+        // Make sure logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    public function manage(Request $request){
+        $listings = auth()->user()->listings;
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get(),
+        ]);
+
     }
 }
